requirements.txt
pandas
sqlalchemy
yfinance
requests
python-dateutil
